// packages/shared/src/enums.ts
export enum UserRole {
  PLAYER = "PLAYER",
  SUPPORT = "SUPPORT",
  KYC_OFFICER = "KYC_OFFICER",
  PAYMENTS = "PAYMENTS",
  RISK = "RISK",
  AFFILIATE_MANAGER = "AFFILIATE_MANAGER",
  MARKETING = "MARKETING",
  ADMIN = "ADMIN",
  SUPER_ADMIN = "SUPER_ADMIN",
}

export const ADMIN_ROLES: UserRole[] = [
  UserRole.SUPPORT,
  UserRole.KYC_OFFICER,
  UserRole.PAYMENTS,
  UserRole.RISK,
  UserRole.AFFILIATE_MANAGER,
  UserRole.MARKETING,
  UserRole.ADMIN,
  UserRole.SUPER_ADMIN,
];